api/requirements.txt
fastapi[standard]
python-dotenv
huggingface_hub
aiofiles
qdrant-client[fastembed]